% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_data.R
\name{available_data}
\alias{available_data}
\title{Print the list available datasets.}
\usage{
available_data()
}
\value{
Nothing.
}
\description{
For each dataset a code and brief descriptions are reported.
}
\examples{
available_data()
}
