% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_CROSS_CRC_ADENOCARCINOMA_NETECOEVO.R
\docType{data}
\name{CROSS_CRC_ADENOCARCINOMA_NATECOEVO_2018}
\alias{CROSS_CRC_ADENOCARCINOMA_NATECOEVO_2018}
\title{Mutations and copy number from colorectal adenocarcinomas (PMID: 30177804).
n = 19 patients, multi-region whole-exome sequencing. Calls generated from
Cross et al. (PMID: 30177804).}
\format{Data for the adenocarcinoma cohort by Cross et al. The data has been provided
by WCh Cross, as of the Nature Ecology and Evolution paper (PMID: 30177804).
Mutations are reported as present or absent in each one of the sequenced
biopsies. All the data (calls, definition of driver events, etc.) is as provided
by the authos, and as used in the paper Caravagna et al (PMID: 30171232) to show
repeated evolution in breast cancers with REVOLVER.}
\usage{
data(CROSS_CRC_ADENOCARCINOMA_NATECOEVO_2018)
}
\description{
Mutations and copy number from colorectal adenocarcinomas (PMID: 30177804).
n = 19 patients, multi-region whole-exome sequencing. Calls generated from
Cross et al. (PMID: 30177804).
}
\examples{
data(CROSS_CRC_ADENOCARCINOMA_NATECOEVO_2018)
CROSS_CRC_ADENOCARCINOMA_NATECOEVO_2018
}
\references{
Cross, W., Kovac, M., Mustonen, V., Temko, D., Davis, H., Baker, A.M.,
Biswas, S., Arnold, R., Chegwidden, L., Gatenbee, C. et al., 2018.
The evolutionary landscape of colorectal tumorigenesis.
Nature Ecology and Evolution, 2(10), p.1661. PMID 30177804.
}
\keyword{datasets}
