% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TRACERx_NEJM_2017_REVOLVER}
\alias{TRACERx_NEJM_2017_REVOLVER}
\title{REVOLVER analysis of dataset TRACERx_NEJM_2017.}
\format{A fit returned from the analysis with the \code{\link{revolver}} package.}
\usage{
data(TRACERx_NEJM_2017_REVOLVER)
}
\description{
REVOLVER analysis of dataset TRACERx_NEJM_2017.
}
\examples{
data(TRACERx_NEJM_2017_REVOLVER)
TRACERx_NEJM_2017_REVOLVER
}
\references{
Jamal-Hanjani et al. (2017) NEJM 376(22), 2109–2121 (2017).
(\href{http://www.nejm.org/doi/full/10.1056/NEJMoa1616288}{NEJM}). PMID: 28445112.
}
\keyword{datasets}
