% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{YATES_BREAST_NATMED_2015}
\alias{YATES_BREAST_NATMED_2015}
\title{Mutations and copy number from breast cancer patients (PMID: 26099045).
n = 50 patients, multi-region whole-exome sequencing. Calls generated from
Yates et al. (PMID: 26099045).}
\format{Data for the breast cancer cohort by Yates et al. The data has been extracted
from the Supplementary tables released in the original Nature Medicine paper (PMID: 26099045).
Mutations and copy number are reported as present or absent in each one of the sequenced
biopsies. All the data (calls, definition of driver events, etc.) is as released in the
original paper. The data is as used in the paper Caravagna et al (PMID: 30171232) to show
repeated evolution in breast cancers with REVOLVER.}
\usage{
data(YATES_BREAST_NATMED_2015)
}
\description{
Mutations and copy number from breast cancer patients (PMID: 26099045).
n = 50 patients, multi-region whole-exome sequencing. Calls generated from
Yates et al. (PMID: 26099045).
}
\examples{
data(YATES_BREAST_NATMED_2015)
YATES_BREAST_NATMED_2015
}
\references{
Yates, L.R., Gerstung, M., Knappskog, S., Desmedt, C., Gundem, G., Van Loo, P.,
Aas, T., Alexandrov, L.B., Larsimont, D., Davies, H. et al., 2015. Subclonal
diversification of primary breast cancer revealed by multiregion sequencing.
Nature Medicine, 21(7), p.751. PMID: 26099045
}
\keyword{datasets}
