% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TRACERx_NEJM_2017}
\alias{TRACERx_NEJM_2017}
\title{Mutations from TRACERx lung (NEJM2017, PMID: 28445112). n = 99 patients, multi-region WES}
\format{Data for the TRACER cohort. Somatic mutations (~65K) from >300 tumour biopsies,
whole-exome sequenced at high-depth (~400x). For every mutation the Cancer Fraction is
reported - VAF adjusted for tumour content and copy number data. The data is as used in
the paper Caravagna et al (PMID: 30171232), and can be used to run REVOLVER.}
\usage{
data(TRACERx_NEJM_2017)
}
\description{
Mutations from TRACERx lung (NEJM2017, PMID: 28445112). n = 99 patients, multi-region WES
}
\examples{
data(TRACERx_NEJM_2017)
TRACERx_NEJM_2017
}
\references{
Jamal-Hanjani et al. (2017) NEJM 376(22), 2109–2121 (2017).
(\href{http://www.nejm.org/doi/full/10.1056/NEJMoa1616288}{NEJM}). PMID: 28445112.
}
\keyword{datasets}
